Use correct order of params and enforce type for VideoCapture

+++ videoio.cpp
@@ -17,7 +17,7 @@ int VideoCapture_OpenDevice(VideoCapture v, int device) {
     return v->open(device);
 }
   
-void VideoCapture_Set(VideoCapture v, int prop, int param) {
+void VideoCapture_Set(VideoCapture v, int prop, double param) {
     v->set(prop, param);
 }
